+++ lua/plugins/config/bufferline_config.lua
@@ -2,7 +2,7 @@ return function(use)
 	use({
 		'akinsho/nvim-bufferline.lua',
 		requires = 'kyazdani42/nvim-web-devicons',
-		event = 'BufRead',
+		event = 'VimEnter',
 		config = function()
 			require('bufferline').setup({
 				custom_filter = function(bufnr)